# ARROW-5935: [C++] ArrayBuilder::type() should be kept accurate