… cpp/src/arrow/array/builder_binary.cc
… cpp/src/arrow/array/builder_binary.h
… cpp/src/arrow/array/builder_decimal.cc
… cpp/src/arrow/array/builder_decimal.h
… cpp/src/arrow/array/builder_dict.h
… cpp/src/arrow/array/builder_nested.cc
… cpp/src/arrow/array/builder_nested.h
… cpp/src/arrow/array/builder_primitive.cc
… cpp/src/arrow/array/builder_primitive.h
… cpp/src/arrow/array/builder_time.h
… cpp/src/arrow/array/builder_union.cc
… cpp/src/arrow/array_dict_test.cc
… cpp/src/arrow/array_list_test.cc
… cpp/src/arrow/array_test.cc
… cpp/src/arrow/array_union_test.cc
… cpp/src/arrow/builder.cc
… cpp/src/arrow/python/python_to_arrow.cc
… cpp/src/arrow/python/serialize.cc
… cpp/src/arrow/type.cc