`ArrayBuilder::type()` should always be identical to the type of the array which would be produced by `ArrayBuilder::Finish()`, even for builders whose type may change. This includes:
- adaptive int builders, which may increase the size of integers built
- dictionary builders, which may increase the size of their indices
- union builders, which may append a new child
- nested builders whose children include a builder with mutable type