@bkietz 2019-07-26
note that this is the current solution for struct builders with mutable type children: https://github.com/apache/arrow/pull/4930/files#diff-1744e8000519050e6464a5c421995df3R358
(all fields are named "")