@bkietz 2019-07-26
An alternative way to approach this problem would be: extend ArrayBuilder with type mutation callbacks. When a nested builder takes ownership of a child builder, it passes this callback to the child. If the child's type mutates then it must invoke the callback, allowing the parent's type to update as well.