@bkietz 2019-07-31
@pitrou please take a look. I've updated the patch to keep `ArrayBuilder::type_` correct instead of just setting it to `nullptr`. Does this seem less problematic?