@bkietz 2019-07-31
Three dictionary builder benchmarks got ~5% slower as a result of this change:

```
BuildInt64DictionaryArraySimilar:    -0.055
BuildInt64DictionaryArrayRandom:     -0.054
BuildInt64DictionaryArraySequential: -0.051
```