@wesm 2019-08-01
Those benchmarks are really sensitive as we found in ARROW-5527