@bkietz 2019-08-06
@pitrou @wesm does this seem reasonable?