@pitrou 2019-08-07
The adaptive int builder also regresses in performance.
Before:
```
---------------------------------------------------------------------------
Benchmark                                    Time           CPU Iterations
---------------------------------------------------------------------------
BuildAdaptiveIntNoNulls               32935759 ns   32806979 ns         85   7.62033GB/s
BuildAdaptiveIntNoNullsScalarAppend  255407613 ns  255338156 ns         11   1002.59MB/s
```
After:
```
---------------------------------------------------------------------------
Benchmark                                    Time           CPU Iterations
---------------------------------------------------------------------------
BuildAdaptiveIntNoNulls               36321264 ns   36119954 ns         78   6.92138GB/s
BuildAdaptiveIntNoNullsScalarAppend  256083900 ns  255858025 ns         11   1000.55MB/s
```