@pitrou 2019-08-07
Since it seems the implementation also adds some complication, perhaps it's better for `type()` to return `nullptr` after all. And/or there can be two different methods:
* `type()` returning the initialization type
* `actual_type()` returning some else or `nullptr`

(or `initial_type()` vs. `type()` perhaps)