@bkietz 2019-08-07
@pitrou having `type_ == nullptr` is not ideal since there is a lot of metadata stored there. (Significantly: the field names of a struct builder).

The adaptive int builders are only used by the dictionary builders, and those have a 5% regression for dictionaries of integers and a 1.8% regression for dictionaries of strings. If that is unacceptable, I can refactor to make UpdateType public (type_ will not be guaranteed accurate unless UpdateType has been called more recently than Append*)