@wesm 2019-08-07
I might be missing something but would making `ArrayBuilder::type` virtual help? It seems like we might be going to a lot of trouble to make `type_` always accurate