@bkietz 2019-08-07
`type()` can be made virtual but then it's going to become way more expensive for deeply nested types. Is that acceptable?