@pitrou 2019-08-07
We should step back: what is `Builder::type()` useful / used for?