@bkietz 2019-08-07
The most prominent use case currently is to access the type id, which is then used as a discriminant in a switch/visit to unpack the builder