@pitrou 2019-08-07
But that's entirely wrong for a `AdaptiveBuilder`, right?