@bkietz 2019-08-07
I suppose so, but if we allow AdaptiveIntBuilder to refuse the bequest of ArrayBuilder::type then what should `dictionary_builder.type()` be?