@pitrou 2019-08-07
Good question. But if only `type_id()` is required then it becomes easier, right?