@bkietz 2019-08-07
Okay, if we drop ArrayBuilder::type() in favor of ArrayBuilder::type_id() that does indeed make things easier. Slight difficulty: some type_ids don't have a single builder implementation (dictionary builder is templated on the value type). We'd have to provide an "entry point" base class for each type_id so that builders which have addition type metadata can be unpacked appropriately. For example, if builder.type_id() == Type::DICTIONARY, you can cast to DictionaryBuilderBase safely. From there, you can check dict_builder.value_type_id() ==Type::STRING and cast to DictionaryBuidler<StringType>