@pitrou 2019-08-07
I'm starting to think that it deserves a discussion on the ML, because there seem to be multiple ways to solve this and each has non-trivial implications.