@bkietz 2019-08-07
Alright, I can write a summary of the difficulties and our discussion of the solutions here