@pitrou 2019-09-10
I see no noticeable difference on builder and CSV benchmarks here.