@bkietz 2019-09-10
@kou I'll follow up on efficient builder visitation in https://issues.apache.org/jira/browse/ARROW-4146