@pitrou 2019-09-10
Unfortunately the benchmarks above were run on the borked rebase, so I don't think they are reliable :-/